feat: modularizar funções em arquivos diferentes
