+++ moduloAssinante.c
@@ -1,0 +1,14 @@
+void telaAssinante(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║     MÓDULO ASSINANTES    ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Cadastrar Assinantes  ║\n");
+    printf("║ 2. Checar Assinantes     ║\n");
+    printf("║ 3. Alterar Assinantes    ║\n");
+    printf("║ 4. Excluir Assinantes    ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+
+}
