+++ moduloAssinaturas.c
@@ -1,0 +1,13 @@
+void telaAssinaturas(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║     MÓDULO ASSINATURAS   ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Cadastrar Assinaturas ║\n");
+    printf("║ 2. Checar Assinaturas    ║\n");
+    printf("║ 3. Alterar Assinaturas   ║\n");
+    printf("║ 4. Excluir Assinaturas   ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+}
