+++ moduloFinanceiro.c
@@ -1,0 +1,13 @@
+void telaFinanceiro(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║     MÓDULO Financeiro    ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Relatório de pagamento║\n");
+    printf("║ 2. Caixa Mensal          ║\n");
+    printf("║ 3. lorem ipsum           ║\n");
+    printf("║ 4. lorem ipsum           ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+}
