+++ moduloPlanos.c
@@ -1,0 +1,13 @@
+void telaPlano(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║       MÓDULO PLANOS      ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Cadastrar Plano       ║\n");
+    printf("║ 2. Checar Planos         ║\n");
+    printf("║ 3. Alterar Plano         ║\n");
+    printf("║ 4. Excluir Plano         ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+}
