+++ moduloProduto.c
@@ -1,0 +1,13 @@
+void telaProdutos(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║     MÓDULO PRODUTOS      ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Cadastrar Produtos    ║\n");
+    printf("║ 2. Checar Produtos       ║\n");
+    printf("║ 3. Alterar Produtos      ║\n");
+    printf("║ 4. Excluir Produtos      ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+}
