+++ moduloRelatorios.c
@@ -1,0 +1,13 @@
+void telaRelatorios(){
+    system("clear||cls");
+    printf("╔══════════════════════════╗\n");
+    printf("║     MÓDULO Relatórios    ║\n");
+    printf("╠══════════════════════════╣\n");
+    printf("║ 1. Faixa etária          ║\n");
+    printf("║ 2. Vizualizar Assinantes ║\n");
+    printf("║ 3. Vizualizar Assinaturas║\n");
+    printf("║ 4. lorem ipsum           ║\n");
+    printf("║ 5. Sair                  ║\n");
+    printf("╚══════════════════════════╝\n");
+    printf("Digite sua escolha: \n");
+}
